Accidental extremely large commit for new worldgen & planet stuff
 - Moves most worldgen to data. What rest is not there yet needs code not done yet
 - Starts process of moving to PlanetProperties and standard MC dimension creation, does not finish
 - Starts process of overhauling stations/etc because of the above changes to DimensionProperties, does not finish
 - Does not currently comiple.
 - Large API rework in the body (Station/Planet/Star) section

+++ src/main/java/zmaster587/advancedRocketry/api/StatsRocket.java
@@ -73,7 +73,7 @@ public class StatsRocket {
 	public int getMass() {return mass;}
 	public float getDrillingPower() {return drillingPower;}
 	public void setDrillingPower(float power) {drillingPower = power;}
-	public float getAcceleration(float gravitationalMultiplier) { return (getThrust() - (mass * ((ARConfiguration.getCurrentConfig().gravityAffectsFuel.get()) ? gravitationalMultiplier : 1)))/10000f; }
+	public double getAcceleration(double gravitationalMultiplier) { return (getThrust() - (mass * ((ARConfiguration.getCurrentConfig().gravityAffectsFuel.get()) ? gravitationalMultiplier : 1)))/10000f; }
 	public float getAcceleration() { return getThrust()/10000f; }
 	public List<Vector3F<Float>> getEngineLocations() { return engineLoc; }
 	public boolean isNuclear() {return workingFluidTank.getBaseFuelRate() > 0;}
